conducts code cleanup of some sandbox classes removes unnecessary duplicates modernizes some code style towards Java 8+ adds missing break statements in loops to terminate earlier simplifies some methods fixes problematic JavaDoc removes unused imports

+++ nlp-utils/src/main/java/org/apache/opennlp/utils/languagemodel/NaiveSentenceLanguageModel.java
@@ -29,7 +29,7 @@ public class NaiveSentenceLanguageModel<T> implements LanguageModel<T[]> {
 
   @Override
   public double calculateProbability(Collection<T[]> vocabulary, T[] sentence) {
-    return vocabulary.isEmpty() ? 0 : Collections.frequency(vocabulary, sentence) / vocabulary.size();
+    return vocabulary.isEmpty() ? 0 : (double) Collections.frequency(vocabulary, sentence) / vocabulary.size();
   }
 
 }
